prompt-strategies/requirements.txt
mistralai>=1.0.0
asyncio
python-dotenv>=1.0.0
